+++ Realization/backend/acc/src/main/java/eu/bcvsolutions/idm/acc/service/impl/ContractSynchronizationExecutor.java
@@ -75,6 +75,8 @@ import eu.bcvsolutions.idm.core.eav.api.entity.FormableEntity;
 import eu.bcvsolutions.idm.core.eav.api.service.FormService;
 import eu.bcvsolutions.idm.core.model.entity.IdmIdentityContract;
 import eu.bcvsolutions.idm.core.model.entity.IdmTreeNode_;
+import eu.bcvsolutions.idm.core.model.event.ContractGuaranteeEvent;
+import eu.bcvsolutions.idm.core.model.event.ContractGuaranteeEvent.ContractGuaranteeEventType;
 import eu.bcvsolutions.idm.core.model.event.IdentityContractEvent;
 import eu.bcvsolutions.idm.core.model.event.IdentityContractEvent.IdentityContractEventType;
 import eu.bcvsolutions.idm.core.workflow.service.WorkflowProcessInstanceService;
@@ -478,8 +480,10 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 
 			// Delete guarantees
 			guaranteesToDelete.forEach(guarantee -> {
-				// TODO: convert to publish event (with skipProvisioning property)
-				guaranteeService.delete(guarantee);
+				EntityEvent<IdmContractGuaranteeDto> guaranteeEvent = new ContractGuaranteeEvent(
+						ContractGuaranteeEventType.DELETE, guarantee,
+						ImmutableMap.of(ProvisioningService.SKIP_PROVISIONING, skipProvisioning));
+				guaranteeService.publish(guaranteeEvent);
 			});
 
 			// Create new guarantees
@@ -487,8 +491,11 @@ public class ContractSynchronizationExecutor extends AbstractSynchronizationExec
 				IdmContractGuaranteeDto guarantee = new IdmContractGuaranteeDto();
 				guarantee.setIdentityContract(contract.getId());
 				guarantee.setGuarantee(identity.getId());
-				// TODO: convert to publish event (with skipProvisioning property)
-				guaranteeService.save(guarantee);
+				//
+				EntityEvent<IdmContractGuaranteeDto> guaranteeEvent = new ContractGuaranteeEvent(
+						ContractGuaranteeEventType.CREATE,
+						guarantee, ImmutableMap.of(ProvisioningService.SKIP_PROVISIONING, skipProvisioning));
+				guaranteeService.publish(guaranteeEvent);
 			});
 		}