+++ Realization/frontend/czechidm-core/src/content/requestrole/RoleRequestTable.js
@@ -154,7 +154,7 @@ class RoleRequestTable extends Advanced.AbstractTableContent {
               ({ rowIndex, data }) => {
                 const entity = data[rowIndex];
                 return (
-                  <Advanced.IdentityInfo entityIdentifier={ entity.applicant } entity={ entity._embedded.applicant } face="link" />
+                  <Advanced.IdentityInfo entityIdentifier={ entity.applicant } entity={ entity._embedded.applicant } face="popover" />
                 );
               }
             }/>